backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
django-cors-headers==4.3.1
python-dotenv==1.0.1
gunicorn==21.2.0
Pillow==10.2.0
torch==2.2.1
transformers==4.37.2
ultralytics==8.1.28
openai==1.12.0
python-jose==3.3.0
django-filter==23.5
drf-yasg==1.21.7
django-redis==5.4.0
hiredis==2.3.2